Make `PollEditor` its own component and add submit functionality to new poll page.
